% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinity-package.R
\docType{package}
\name{affinity-package}
\alias{affinity}
\alias{affinity-package}
\title{affinity: Basic Tools for Raster Georeferencing, Grid Affine Transforms, and Raster Cell Abstractions}
\description{
Tools for raster georeferencing, grid affine transforms, and general raster logic.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/affinity}
  \item Report bugs at \url{https://github.com/hypertidy/affinity/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
