% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinething.R
\name{assignproj}
\alias{assignproj}
\title{Assign projection}
\usage{
assignproj(x, proj = "+proj=longlat +datum=WGS84")
}
\arguments{
\item{x}{spatial object for use with \code{\link[raster:projection]{raster::projection()}}}

\item{proj}{PROJ.4 string}
}
\value{
a spatial object with the projection set
}
\description{
Set the projection of a spatial object.
}
