% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{extent_dim_to_gt}
\alias{extent_dim_to_gt}
\title{Create geotransform from extent and dimension}
\usage{
extent_dim_to_gt(x, dim)
}
\arguments{
\item{x}{extent parameters, c(xmin,xmax,ymin,ymax)}

\item{dim}{dimensions x,y of grid (ncol,nrow)}
}
\value{
6-element \code{\link[=geo_transform0]{geo_transform0()}}
}
\description{
Create the geotransform (see \code{\link[=geo_transform0]{geo_transform0()}}) from extent and dimension.
}
\details{
The dimension is always ncol, nrow.
}
\examples{
extent_dim_to_gt(c(0, 5, 0, 10), c(5, 10))
}
