\name{getCoreInstances}
\alias{getCoreInstances}
\alias{printCore}

\title{Core instances}
\description{
	The core instances of a category
}
\usage{
getCoreInstances(categoriesId, count = 5)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{categoriesId}{The category for which to print the core instances}
  \item{count}{How many instances to print}
}
\details{
	Core instances for a category are those with the highest centrality values for that category.
}
\value{
}
\author{Claudio Baccigalupo}
\examples{
# getCoreInstances(1)
}

\keyword{methods}
