% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinity-package.R
\docType{data}
\name{monterey}
\alias{monterey}
\title{Monterey Bay elevation}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 270 rows and 270 columns.
}
\usage{
monterey
}
\description{
Extent is in the examples, stolen from rayshader.
}
\details{
A matrix 540x540 of topography. Used in [affinething()] examples.
}
\examples{
ex <- c(-122.366765, -121.366765, 36.179392, 37.179392)
}
\keyword{datasets}
