\name{mystrandsArtistsCooccurrences}
\alias{mystrandsArtistsCooccurrences}
\alias{mystrandsArtistsGenres}
\alias{mystrandsArtistsTags}
\alias{mystrandsTagsNames}
\docType{data}
\title{MyStrands dataset}
\description{
	Dataset of MyStrands artists/genres/tags co-occurrences
}
\usage{data(mystrandsArtistsCooccurrences)}
\format{
  Data frames with MyStrands dataset
}
\details{
	A sample dataset containing a list of 4,000 artists, of 26 genres and of co-occurrences of songs by these artists and genres in the playlists of the users of MyStrands.
}
\source{
	http://www.mystrands.com
}
\examples{
# data(mystrandsArtistsCooccurrences)
}
\keyword{datasets}
