\name{plotAffinity3D}
\alias{plotAffinity3D}

\title{Compare affinity to categories}
\description{
	A 3D plot comparing affinity to 3 categories
}
\usage{
plotAffinity3D(instancesId, threeCategoriesId)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{instancesId}{Instances to plot}
  \item{threeCategoriesId}{Categories to plot}
}
\details{
	This function plots a 3D graph with three categories and the axis, and the fuzzy affinity values of instances to these three categories.
}
\value{
}
\author{Claudio Baccigalupo}
\examples{
# plotAffinity3D(TRUE,1:3)
}

\keyword{methods}
