\name{plotExamples}
\alias{runExamples}
\alias{plotExamples}
\title{Examples}
\description{
	Some textual and plotting example of affinity
}
\usage{
runExamples()
plotExamples()
}

\details{
Once a dataset has been loaded, use these functions to run some examples on random instances and categories.
}
\value{
}
\author{Claudio Baccigalupo}
\examples{
# runExamples()
# plotExamples()
}

\keyword{methods}
