% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{raster_to_gt}
\alias{raster_to_gt}
\title{Geotransform from raster object}
\usage{
raster_to_gt(x)
}
\arguments{
\item{x}{raster object (the raster package, extends BasicRaster)}
}
\value{
a geotransform vector
}
\description{
Return the geotransform defining the raster's offset and resolution.
}
\details{
The geotransform vector is six coefficients xmin, xres, yskew, ymax, xskew, yres, values
relative to the top left corner of the top left pixel. "yres" the y-spacing is
traditionally negative.
}
\examples{
raster_to_gt(raster::raster(volcano))
}
