% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{raster_to_rasterio}
\alias{raster_to_rasterio}
\alias{raster_to_sfio}
\title{RasterIO window from raster object}
\usage{
raster_to_rasterio(x)

raster_to_sfio(x)
}
\arguments{
\item{x}{a raster object (BasicRaster, from raster package)}
}
\value{
RasterIO window vector 'c(x0, y0, nx0, ny0, nx, y)' see Details
}
\description{
Return the RasterIO window vector defining the raster's offset and resolution and dimensions.
}
\details{
The RasterIO window is a six element vector of offset (x,y),  dimension of source (nx0, ny0) and
dimension of output (nx, ny).

The sf RasterIO is the RasterIO window in a list format used by the sf package, it contains the same
information, and is created by \code{\link[=raster_to_sfio]{raster_to_sfio()}}.
}
\examples{
raster_to_rasterio(raster::raster(volcano))
}
