% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{raster_to_world}
\alias{raster_to_world}
\title{World vector from raster object.}
\usage{
raster_to_world(x)
}
\arguments{
\item{x}{raster object (the raster package, extends BasicRaster)}
}
\value{
a geotransform vector
}
\description{
Return the world transform defining the raster's offset and resolution.
}
\details{
The world vector is the values xres, yres, xmin, ymax relative to the
centre of the top left pixel. "yres" the y-spacing is traditionally negative.
}
\examples{
raster_to_world(raster::raster(volcano))
}
