% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{rasterio_to_sfio}
\alias{rasterio_to_sfio}
\title{The sf/stars RasterIO list}
\usage{
rasterio_to_sfio(x)
}
\arguments{
\item{x}{rasterio params as from \code{\link[=raster_io0]{raster_io0()}}}
}
\value{
list in sf RasterIO format
}
\description{
We create the list as used by the stars/sf GDAL IO function 'gdal_read(, RasterIO_parameters)'.
}
\details{
Note that the input is a 4 or 6 element vector, with offset 0-based and
output dimensions optional (will use the source window). The resample argument uses the syntax
identical to that used in GDAL itself.
}
\examples{
rio <- raster_io0(c(0L, 0L), src_dim = c(24L, 10L))
rasterio_to_sfio(rio)
}
