% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinething.R
\name{regions}
\alias{regions}
\title{regions (this should be a service/repo to get stored auxiliaries)}
\usage{
regions(name)
}
\arguments{
\item{name}{}
}
\value{

}
\description{
regions (this should be a service/repo to get stored auxiliaries)
}

