% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{sfio_to_rasterio}
\alias{sfio_to_rasterio}
\title{sf package RasterIO from RasterIO window vector}
\usage{
sfio_to_rasterio(x)
}
\arguments{
\item{x}{a RasterIO parameter list}
}
\value{
a sf-RasterIO parameter list
}
\description{
Basic function to create the window parameters as used by GDAL RasterIO, in
format used by sf, in 'gdal_read(,RasterIO_parameters)'.
}
\examples{
sfio_to_rasterio(rasterio_to_sfio(raster_io0(c(0L, 0L), src_dim = c(24L, 10L))))
}
