% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinething.R
\name{affinething}
\alias{affinething}
\title{affinething}
\usage{
affinething(x, rgb = FALSE)
}
\arguments{
\item{x}{a raster}

\item{rgb}{use RGB plot for a raster with 3 layers}
}
\description{
Obtain control points for the simple affine transform (offset and scale) on an ungeoreferenced image.
}
