\name{affinity}
\alias{affinity}
\docType{package}
\title{
The R affinity package
}
\description{
R functions for printing and plotting fuzzy affinity
}
\details{
This package contains functions to analyze and plot the affinity to multiple categories of an instance based on observed data about the co-occurrences of that instance with other instances of different categories.
}
\author{ Claudio Baccigalupo \email{claudio@iiia.csic.es} }

\references{
Baccigalupo C., Donaldson J. and Plaza E. (2008)
Uncovering affinity of artists to multiple genres from social behaviour data
International Symposium on Music Information Retrieval (ISMIR '08)
}

\keyword{ package }
\seealso{
}
\examples{
# data(mystrandsGenres)
# runExamples()
# plotExamples()
}
