% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinity-package.R
\docType{package}
\name{affinity-package}
\alias{affinity}
\alias{affinity-package}
\title{affinity: Raster Georeferencing, Grid Affine Transforms, Cell Abstraction}
\description{
Tools for raster georeferencing, grid affine transforms, and general raster logic. 
 These functions provide converters between raster specifications, world vector, geotransform, 
 'RasterIO' window, and 'RasterIO window' in 'sf' package list format. There are functions to offset
 a matrix by padding any of four corners (useful for vectorizing neighbourhood operations), and
 helper functions to harvesting user clicks on a graphics device to use for simple georeferencing
 of images.  Methods used are available from <https://en.wikipedia.org/wiki/World_file> and
 <https://gdal.org/user/raster_data_model.html>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/affinity}
  \item Report bugs at \url{https://github.com/hypertidy/affinity/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
