% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinething.R
\name{assignproj}
\alias{assignproj}
\title{assignproj}
\usage{
assignproj(x,
  proj = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0")
}
\arguments{
\item{proj}{}
}
\description{
assignproj
}

