% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinething.R
\name{domath}
\alias{domath}
\title{Calculate the math of an affine transform}
\usage{
domath(pts, xy, r = NULL, proj = NULL)
}
\arguments{
\item{pts}{known points of 'xy'}

\item{xy}{'xy' obtain from `affinething`}

\item{r}{raster in use}

\item{proj}{optional projection, if the pts are longlat and the raster is not}
}
\value{
raster extent
}
\description{
Given relative location and absolute locations, convert to an actual real world extent
for a matrix of data.
}
\details{
Convert known geographic points with raw graphic control points and a reference raster
to an extent for the raster in geography.
}
\seealso{
affinething
}
