% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{geo_transform0}
\alias{geo_transform0}
\title{Geo transform parameter creator}
\usage{
geo_transform0(px, ul, sh = c(0, 0))
}
\arguments{
\item{px}{pixel resolution (XY, Y-negative)}

\item{ul}{grid offset, top-left corner}

\item{sh}{affine shear (XY)}
}
\value{
vector of parameters xmin, xres, yskew, ymax, xskew, yres
}
\description{
Basic function to create a geotransform as used by GDAL.
}
\examples{
geo_transform0(px = c(1, -1), ul = c(0, 0))
}
\seealso{
\code{\link[=geo_world0]{geo_world0()}} which uses the same parameters in a different order
}
